refactor(storage): mark the barrier modules internal and remove the init ordering hazard
Both findings valid.

**Internal boundary.** `store-control-barrier.ts` and `store-barrier-contract.ts`
are renamed to `…-v1-internal.ts`, matching this package's existing convention
(`managed-oxigraph-ownership-v1-internal.ts`). Neither is exported from the
barrel, but the package publishes `dist` wholesale, so the file name is the only
signal a deep importer gets — and `StoreBarrierHostV1` is a private callback
contract between two scheduler internals, not something to couple to. The stable
public types remain the ones re-exported from `store-priority-scheduler.ts`.

**Initialization ordering.** The coordinator was constructed twenty lines before
`this.now = resolvedNow`, so its host closure `now: () => this.now()` would have
read `undefined` had it been called during construction. It is not called today —
the coordinator constructor only stores its arguments — so this was a latent
hazard rather than a live defect, but that is exactly the reviewer's point: the
boundary silently depended on an undocumented property of the other side. A
coordinator that later snapshots a timestamp at construction would have found a
partially built scheduler.

Construction now happens after its dependencies are assigned, so correctness no
longer rests on the callee staying passive.

Storage lane 30 files / 525 tests, typecheck clean.
